Re-organized folder structure and autoload.
