+++ hooks/index.php
@@ -1,0 +1,5 @@
+<?php
+
+/**
+ * Silence is golden.
+ */
